requirements.txt
aniso8601==7.0.0
appnope==0.1.2
argon2-cffi==20.1.0
asgiref==3.3.1
astroid==2.4.2
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
clang==11.0
colorama==0.4.4
coverage==5.5
cycler==0.10.0
Cython==0.29.22
decorator==5.0.5
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.1.6
django-appconf==1.0.4
django-cors-headers==3.7.0
django-filter==2.4.0
django-haystack==3.0
django-heroku==0.0.0
django-image-cropping==1.5.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
drf-haystack==1.8.10
easy-thumbnails==2.7.1
elasticsearch==7.13.4
entrypoints==0.3
geographiclib==1.52
geopy==2.2.0
google-streetview==1.2.9
googlemaps==4.4.5
graphene==2.1.8
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.1.0
idna==2.10
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==5.7.0
jedi==0.18.0
jellyfish==0.6.1
Jinja2==2.11.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.13
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.1.0
kwconfig==1.1.7
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.4.1
mccabe==0.6.1
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.3.0
numpy==1.20.2
packaging==20.9
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pi==0.1.2
pickleshare==0.7.5
pigar==0.10.0
Pillow==8.2.0
prometheus-client==0.10.0
promise==2.3
prompt-toolkit==3.0.18
ptyprocess==0.7.0
pycparser==2.20
pygame==2.0.1
Pygments==2.8.1
PyJWT==1.7.1
pylint==2.6.0
pyparsing==2.4.7
pyrsistent==0.17.3
pysolr==3.9.0
python-dateutil==2.8.1
pytz==2021.1
pyzmq==22.0.3
qtconsole==5.0.3
QtPy==1.9.0
requests==2.25.1
Rx==1.6.1
Send2Trash==1.5.0
singledispatch==3.6.1
six==1.15.0
sqlparse==0.4.1
terminado==0.9.4
testpath==0.4.4
text-unidecode==1.3
toml==0.10.2
tornado==6.1
traitlets==5.0.5
urllib3==1.26.4
us==2.0.2
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.2.0
Whoosh==2.7.4
widgetsnbextension==3.5.1
wrapt==1.12.1
